Move `RedisConnector` from `components/` to `server/`
