+++ packages/server/src/RedisConnector.ts
@@ -1,6 +1,6 @@
-import { InternalFlowiseError } from '../../server/src/errors/internalFlowiseError'
-import logger from '../../server/src/utils/logger'
-import { MODE } from '../../server/src/Interface'
+import { InternalFlowiseError } from './errors/internalFlowiseError'
+import logger from './utils/logger'
+import { MODE } from './Interface'
 import { Redis } from 'ioredis'
 import { StatusCodes } from 'http-status-codes'
 
@@ -34,7 +34,7 @@ export class RedisConnector {
 
     /**
      * Sync constructor
-     * 
+     *
      * @constructor
      */
     constructor() {}
